revisit fast diff

+++ scratch/kernels/diff.c
@@ -1,5 +1,5 @@
 #include <stdint.h>
-void func(short* restrict s, short* restrict t, int n) {
+void func(short* restrict s, int* restrict t, int n) {
   for (int i = 0 ; i < n; i +=2) {
     t[i/2] = s[i] - s[i + 1];
   }
